Refactor tasklist-class.js to move renderList method to a separate file, further refactor renderList method to move updateInput, updateStatus, deleteItem functions to separate files
